created Cacheable interface CacheDataStore now adds new interface to all created proxies, new method to clearCache without closing datastore.  fixed compile warnings in various classes.
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@381 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/trace/sanger/traceArchive/DefaultTraceNameConverter.java
@@ -48,7 +48,7 @@ public class DefaultTraceNameConverter implements IdConverter<String,Long> {
      * @throws DataStoreException if there is a problem reading the traceInfo.
      */
     public DefaultTraceNameConverter(TraceArchiveInfo traceInfo, TraceArchiveRecordIdGenerator traceIdGenerator ) throws DataStoreException {
-        map = new HashMap<String, Long>((int)traceInfo.size());
+        map = new HashMap<String, Long>(traceInfo.size());
         for(TraceArchiveRecord record : traceInfo){
             String traceId = traceIdGenerator.generateIdFor(record);
             Long traceName = Long.parseLong(record.getAttribute(TraceInfoField.TRACE_NAME));
